+++ sakura_core/mem/CMemory.h
@@ -44,7 +44,7 @@ class CMemory
 	//コンストラクタ・デストラクタ
 public:
 	CMemory() noexcept;
-	CMemory(const void* pData, int nDataLenBytes);
+	CMemory(const void* pData, SSIZE_T nDataLenBytes);
 	CMemory(const CMemory& rhs);
 	CMemory(CMemory&& other) noexcept;
 	// デストラクタを仮想にすると仮想関数テーブルへのポインタを持つ為にインスタンスの容量が増えてしまうので仮想にしない
@@ -54,11 +54,11 @@ public:
 	//インターフェース
 public:
 	void AllocBuffer(SSIZE_T nNewDataLen);                               //!< バッファサイズの調整。必要に応じて拡大する。
-	void SetRawData( const void* pData, int nDataLen );    //!< バッファの内容を置き換える
+	void SetRawData( const void* pData, SSIZE_T nDataLen );    //!< バッファの内容を置き換える
 	void SetRawData(const CMemory& pcmemData);                     //!< バッファの内容を置き換える
-	void SetRawDataHoldBuffer( const void* pData, int nDataLen );    //!< バッファの内容を置き換える(バッファを保持)
+	void SetRawDataHoldBuffer( const void* pData, SSIZE_T nDataLen );    //!< バッファの内容を置き換える(バッファを保持)
 	void SetRawDataHoldBuffer(const CMemory& pcmemData);                     //!< バッファの内容を置き換える(バッファを保持)
-	void AppendRawData( const void* pData, int nDataLen ); //!< バッファの最後にデータを追加する
+	void AppendRawData( const void* pData, SSIZE_T nDataLen ); //!< バッファの最後にデータを追加する
 	void AppendRawData(const CMemory* pcmemData);                  //!< バッファの最後にデータを追加する
 	void Clean(){ _Empty(); }
 	void Clear(){ _Empty(); }
@@ -88,7 +88,7 @@ public:
 	static int IsEqual(const CMemory& cmem1, const CMemory& cmem2);	/* 等しい内容か */
 
 	// 変換関数
-	static void SwapHLByte(char* pData, const int nDataLen); // 下記関数のstatic関数版
+	static void SwapHLByte(char* pData, const SSIZE_T nDataLen); // 下記関数のstatic関数版
 	void SwapHLByte();			// Byteを交換する
 	bool SwabHLByte(const CMemory& mem); // Byteを交換する(コピー版)
 
@@ -97,16 +97,16 @@ protected:
 	||  実装ヘルパ関数
 	*/
 	void _Empty( void ); //!< 解放する。m_pRawDataはNULLになる。
-	void _AddData(const void* pData, int nDataLen);
+	void _AddData(const void* pData, SSIZE_T nDataLen);
 public:
 	void _AppendSz(const char* str);
-	void _SetRawLength(int nLength);
+	void _SetRawLength(SSIZE_T nLength);
 	void swap( CMemory& left ) noexcept {
 		std::swap( m_nDataBufSize, left.m_nDataBufSize );
 		std::swap( m_pRawData, left.m_pRawData );
 		std::swap( m_nRawLen, left.m_nRawLen );
 	}
-	int capacity() const { return m_nDataBufSize ? m_nDataBufSize - 2: 0; }
+	SSIZE_T capacity() const { return m_nDataBufSize ? m_nDataBufSize - 2: 0; }
 
 private: // 2002/2/10 aroka アクセス権変更
 	/*
